requirements.txt
# Billboard Detection Model Training Requirements

# Core training framework
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Data processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Annotation tools (optional - for manual labeling)
labelImg>=1.8.6

# Export and utilities
onnx>=1.14.0
onnxruntime>=1.16.0
